Update v2.1.3
added feature #1

State content can now be added in the bar itself.
State content in the bar is separate from the legend.
Added percentage [%] option to both state contents.
Font size of the bar state content can be adjusted.
Position of the bar state content can be adjusted.
Text color is adjusted based on the color of the bar.

Position changed to "center" instead of "middle" (vertical / horizontal - consistency and clarity)

The unit of the main entity is now smaller and gray.

Tried to fix the interactions - kind of succeeded
Also it was driving me mad that i couldn't right click on the card xd

Documentation will be in the next commit

General cleanup.